fix: WebSocket과 관련된 채팅 컨트롤러 및 서비스 분리
